requirements.txt
flask  
gradio  
pillow  
matplotlib  
ultralytics  
numpy  
pandas  